Init and move MosaicDetection
