+++ src/netspresso_trainer/dataloaders/augmentation/custom.py
@@ -24,146 +24,6 @@ INVERSE_MODES_MAPPING = {
 }
 
 
-def adjust_box_anns(bbox, scale_ratio, padw, padh, w_max, h_max):
-    bbox[:, 0::2] = np.clip(bbox[:, 0::2] * scale_ratio + padw, 0, w_max)
-    bbox[:, 1::2] = np.clip(bbox[:, 1::2] * scale_ratio + padh, 0, h_max)
-    return bbox
-
-
-def getRotationMatrix2D(angle, center, scale):
-    angle = np.deg2rad(angle)
-    alpha = scale * np.cos(angle)
-    beta = scale * np.sin(angle)
-    mat = [
-        [alpha, beta, (1-alpha)*center[0] - beta*center[1]],
-        [-beta, alpha, beta*center[0] + (1-alpha)*center[1]]
-    ]
-    mat = np.array(mat)
-    return mat
-
-
-def get_mosaic_coordinate(mosaic_image, mosaic_index, xc, yc, w, h, input_h, input_w):
-    # TODO update doc
-    # index0 to top left part of image
-    if mosaic_index == 0:
-        x1, y1, x2, y2 = max(xc - w, 0), max(yc - h, 0), xc, yc
-        small_coord = w - (x2 - x1), h - (y2 - y1), w, h
-    # index1 to top right part of image
-    elif mosaic_index == 1:
-        x1, y1, x2, y2 = xc, max(yc - h, 0), min(xc + w, input_w * 2), yc
-        small_coord = 0, h - (y2 - y1), min(w, x2 - x1), h
-    # index2 to bottom left part of image
-    elif mosaic_index == 2:
-        x1, y1, x2, y2 = max(xc - w, 0), yc, xc, min(input_h * 2, yc + h)
-        small_coord = w - (x2 - x1), 0, w, min(y2 - y1, h)
-    # index2 to bottom right part of image
-    elif mosaic_index == 3:
-        x1, y1, x2, y2 = xc, yc, min(xc + w, input_w * 2), min(input_h * 2, yc + h)  # noqa
-        small_coord = 0, 0, min(w, x2 - x1), min(y2 - y1, h)
-    return (x1, y1, x2, y2), small_coord
-
-
-def get_aug_params(value, center=0):
-    if isinstance(value, float):
-        return random.uniform(center - value, center + value)
-    elif len(value) == 2:
-        return random.uniform(value[0], value[1])
-    else:
-        raise ValueError(
-            "Affine params should be either a sequence containing two values\
-             or single float values. Got {}".format(value)
-        )
-
-
-def get_affine_matrix(
-    target_size,
-    degrees=10,
-    translate=0.1,
-    scales=0.1,
-    shear=10,
-):
-    twidth, theight = target_size
-
-    # Rotation and Scale
-    angle = get_aug_params(degrees)
-    scale = get_aug_params(scales, center=1.0)
-
-    if scale <= 0.0:
-        raise ValueError("Argument scale should be positive")
-
-    R = getRotationMatrix2D(angle=angle, center=(0, 0), scale=scale)
-
-    M = np.ones([2, 3])
-    # Shear
-    shear_x = math.tan(get_aug_params(shear) * math.pi / 180)
-    shear_y = math.tan(get_aug_params(shear) * math.pi / 180)
-
-    M[0] = R[0] + shear_y * R[1]
-    M[1] = R[1] + shear_x * R[0]
-
-    # Translation
-    translation_x = get_aug_params(translate) * twidth  # x translation (pixels)
-    translation_y = get_aug_params(translate) * theight  # y translation (pixels)
-
-    M[0, 2] = translation_x
-    M[1, 2] = translation_y
-
-    return M, scale
-
-
-def apply_affine_to_bboxes(targets, target_size, M, scale):
-    num_gts = len(targets)
-
-    # warp corner points
-    twidth, theight = target_size
-    corner_points = np.ones((4 * num_gts, 3))
-    corner_points[:, :2] = targets[:, [0, 1, 2, 3, 0, 3, 2, 1]].reshape(
-        4 * num_gts, 2
-    )  # x1y1, x2y2, x1y2, x2y1
-    corner_points = corner_points @ M.T  # apply affine transform
-    corner_points = corner_points.reshape(num_gts, 8)
-
-    # create new boxes
-    corner_xs = corner_points[:, 0::2]
-    corner_ys = corner_points[:, 1::2]
-    new_bboxes = (
-        np.concatenate(
-            (corner_xs.min(1), corner_ys.min(1), corner_xs.max(1), corner_ys.max(1))
-        )
-        .reshape(4, num_gts)
-        .T
-    )
-
-    # clip boxes
-    new_bboxes[:, 0::2] = new_bboxes[:, 0::2].clip(0, twidth)
-    new_bboxes[:, 1::2] = new_bboxes[:, 1::2].clip(0, theight)
-
-    targets[:, :4] = new_bboxes
-
-    return targets
-
-
-def random_affine(
-    img,
-    targets=(),
-    target_size=(640, 640),
-    degrees=10,
-    translate=0.1,
-    scales=0.1,
-    shear=10,
-):
-    M, scale = get_affine_matrix(target_size, degrees, translate, scales, shear)
-
-    # cv2 function should be replaced
-    img = cv2.warpAffine(img, M, dsize=target_size, borderValue=(114, 114, 114))
-
-    # Transform label coordinates
-    if len(targets) > 0:
-        targets = apply_affine_to_bboxes(targets, target_size, M, scale)
-
-    return img, targets
-
-
 class Compose:
     def __init__(self, transforms, additional_targets: Dict = None):
         if additional_targets is None:
@@ -937,193 +797,6 @@ class RandomCutmix:
         return "{}(num_classes={}, p={}, alpha={}, inplace={})".format(
             self.__class__.__name__, self.num_classes, self.p, self.alpha, self.inplace
         )
-
-
-class MosaicDetection:
-    """
-    Based on the MosaicDetection implementation of Megvii.
-    https://github.com/Megvii-BaseDetection/YOLOX
-    """
-    visualize = False
-
-    def __init__(
-        self,
-        mosaic_scale: List,
-        mixup_scale: List,
-        degrees: float,
-        translate: float,
-        shear: float,
-        enable_mixup: bool,
-        mosaic_prob: float,
-        mixup_prob: float,
-    ):
-        self.mosaic_scale = mosaic_scale
-        self.mixup_scale = mixup_scale
-        self.degrees = degrees
-        self.translate = translate
-        self.shear = shear
-        self.enable_mixup = enable_mixup
-        self.mosaic_prob = mosaic_prob
-        self.mixup_prob = mixup_prob
-
-        self.enable_mosaic = True
-
-    def __call__(self, image, label=None, mask=None, bbox=None, dataset=None):
-        if self.enable_mosaic and random.random() < self.mosaic_prob:
-            mosaic_labels = []
-            input_dim = (640, 640)
-            input_h, input_w = input_dim[0], input_dim[1]
-
-            # yc, xc = s, s  # mosaic center x, y
-            yc = int(random.uniform(0.5 * input_h, 1.5 * input_h))
-            xc = int(random.uniform(0.5 * input_w, 1.5 * input_w))
-
-            # 3 additional image indices
-            items = [dataset.pull_item(random.randint(0, len(dataset) - 1)) for _ in range(3)]
-            items = [(image, label, bbox)] + items
-
-            c = len(image.split())
-            mosaic_img = np.full((input_h * 2, input_w * 2, c), 114, dtype=np.uint8)
-
-            for i_mosaic, (image, label, bbox) in enumerate(items):
-                #h0, w0 = image.shape[:2]  # orig hw
-                w, h = image.size
-                scale = min(1. * input_h / h, 1. * input_w / w)
-
-                image = F.resize(image, (int(h * scale), int(w * scale)), InterpolationMode.BILINEAR)
-
-                # @illian01: PIL -> ndarray, process with ndarray to modify code little
-                image = np.array(image)
-
-                h, w, _ = image.shape[:3]
-                # generate output mosaic image
-
-                # suffix l means large image, while s means small image in mosaic aug.
-                (l_x1, l_y1, l_x2, l_y2), (s_x1, s_y1, s_x2, s_y2) = get_mosaic_coordinate(
-                    mosaic_img, i_mosaic, xc, yc, w, h, input_h, input_w
-                )
-
-                mosaic_img[l_y1:l_y2, l_x1:l_x2] = image[s_y1:s_y2, s_x1:s_x2]
-                padw, padh = l_x1 - s_x1, l_y1 - s_y1
-
-                m_labels = np.concatenate([bbox, label], axis=-1)
-                # Normalized xywh to pixel xyxy format
-                if m_labels.size > 0:
-                    m_labels[:, 0] = scale * m_labels[:, 0] + padw
-                    m_labels[:, 1] = scale * m_labels[:, 1] + padh
-                    m_labels[:, 2] = scale * m_labels[:, 2] + padw
-                    m_labels[:, 3] = scale * m_labels[:, 3] + padh
-                mosaic_labels.append(m_labels)
-
-            if len(mosaic_labels):
-                mosaic_labels = np.concatenate(mosaic_labels, 0)
-                np.clip(mosaic_labels[:, 0], 0, 2 * input_w, out=mosaic_labels[:, 0])
-                np.clip(mosaic_labels[:, 1], 0, 2 * input_h, out=mosaic_labels[:, 1])
-                np.clip(mosaic_labels[:, 2], 0, 2 * input_w, out=mosaic_labels[:, 2])
-                np.clip(mosaic_labels[:, 3], 0, 2 * input_h, out=mosaic_labels[:, 3])
-
-            mosaic_img, mosaic_labels = random_affine(
-                mosaic_img,
-                mosaic_labels,
-                target_size=(input_w, input_h),
-                degrees=self.degrees,
-                translate=self.translate,
-                scales=self.mosaic_scale,
-                shear=self.shear,
-            )
-
-            # -----------------------------------------------------------------
-            # CopyPaste: https://arxiv.org/abs/2012.07177
-            # -----------------------------------------------------------------
-            if (
-                self.enable_mixup
-                and len(mosaic_labels) != 0
-                and random.random() < self.mixup_prob
-            ):
-                mosaic_img, mosaic_labels = self.mixup(mosaic_img, mosaic_labels, input_dim, dataset)
-
-            bbox = mosaic_labels[:, :4]
-            label = mosaic_labels[:, -1:]
-            return mosaic_img, label, mask, bbox
-
-        else:
-            return image, label, mask, bbox
-
-    def mixup(self, origin_img, origin_labels, input_dim, dataset):
-        jit_factor = random.uniform(*self.mixup_scale)
-        FLIP = random.uniform(0, 1) > 0.5
-        cp_labels = []
-        while len(cp_labels) == 0:
-            cp_index = random.randint(0, len(dataset) - 1)
-            _, cp_labels, _ = dataset.pull_item(cp_index)
-        img, cp_labels, cp_boxes = dataset.pull_item(cp_index)
-        img = np.array(img)
-        cp_labels = np.concatenate([cp_boxes, cp_labels], axis=-1)
-
-        if len(img.shape) == 3:
-            cp_img = np.ones((input_dim[0], input_dim[1], 3), dtype=np.uint8) * 114
-        else:
-            cp_img = np.ones(input_dim, dtype=np.uint8) * 114
-
-        cp_scale_ratio = min(input_dim[0] / img.shape[0], input_dim[1] / img.shape[1])
-        resized_img = cv2.resize(
-            img,
-            (int(img.shape[1] * cp_scale_ratio), int(img.shape[0] * cp_scale_ratio)),
-            interpolation=cv2.INTER_LINEAR,
-        )
-
-        cp_img[
-            : int(img.shape[0] * cp_scale_ratio), : int(img.shape[1] * cp_scale_ratio)
-        ] = resized_img
-
-        cp_img = cv2.resize(
-            cp_img,
-            (int(cp_img.shape[1] * jit_factor), int(cp_img.shape[0] * jit_factor)),
-        )
-        cp_scale_ratio *= jit_factor
-
-        if FLIP:
-            cp_img = cp_img[:, ::-1, :]
-
-        origin_h, origin_w = cp_img.shape[:2]
-        target_h, target_w = origin_img.shape[:2]
-        padded_img = np.zeros(
-            (max(origin_h, target_h), max(origin_w, target_w), 3), dtype=np.uint8
-        )
-        padded_img[:origin_h, :origin_w] = cp_img
-
-        x_offset, y_offset = 0, 0
-        if padded_img.shape[0] > target_h:
-            y_offset = random.randint(0, padded_img.shape[0] - target_h - 1)
-        if padded_img.shape[1] > target_w:
-            x_offset = random.randint(0, padded_img.shape[1] - target_w - 1)
-        padded_cropped_img = padded_img[
-            y_offset: y_offset + target_h, x_offset: x_offset + target_w
-        ]
-
-        cp_bboxes_origin_np = adjust_box_anns(
-            cp_labels[:, :4].copy(), cp_scale_ratio, 0, 0, origin_w, origin_h
-        )
-        if FLIP:
-            cp_bboxes_origin_np[:, 0::2] = (
-                origin_w - cp_bboxes_origin_np[:, 0::2][:, ::-1]
-            )
-        cp_bboxes_transformed_np = cp_bboxes_origin_np.copy()
-        cp_bboxes_transformed_np[:, 0::2] = np.clip(
-            cp_bboxes_transformed_np[:, 0::2] - x_offset, 0, target_w
-        )
-        cp_bboxes_transformed_np[:, 1::2] = np.clip(
-            cp_bboxes_transformed_np[:, 1::2] - y_offset, 0, target_h
-        )
-
-        cls_labels = cp_labels[:, 4:5].copy()
-        box_labels = cp_bboxes_transformed_np
-        labels = np.hstack((box_labels, cls_labels))
-        origin_labels = np.vstack((origin_labels, labels))
-        origin_img = origin_img.astype(np.float32)
-        origin_img = 0.5 * origin_img + 0.5 * padded_cropped_img.astype(np.float32)
-
-        return origin_img.astype(np.uint8), origin_labels
 
 
 class Normalize:
